Move some performance tests to experiment stage
Testing clean builds is very time-consuming and
rarely tells us anything about Gradle's performance,
because it is mostly constrained by the performance
of the tools we are calling.

The first-use use case takes a long time as well and
is not important enough to hold up the commit pipeline.
